tests/submission/W3C/domhtml/HTMLLinkElement08.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>link</title>
<link href="/resources/testharness.css" hreflang="en" media="screen" rev="stylesheet" type="text/css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-link-element.html#dom-link-type" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<p>The type attribute specifies the advisory content type. 

    Retrieve the type attribute and examine its value.</p>

<script>
var domtest = async_test("The type attribute specifies the advisory content type.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtype;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("link");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(0);
      vtype = testNode.type;

      assert_equals(vtype,"text/css");
       
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
